backend/requirements.txt
fastapi
uvicorn[standard]
pillow
python-multipart
pydantic
pytesseract
